# Optimize UriParser component parsing