--- a/core-common/src/main/java/org/glassfish/jersey/uri/internal/UriParser.java
+++ b/core-common/src/main/java/org/glassfish/jersey/uri/internal/UriParser.java
@@ -119,8 +119,8 @@ class UriParser {
                 sb.append(c);
 
                 // test IPv6 or regular expressions in the template params
-            } else if ((delimiters != null && delimiters.indexOf(c) >= 0)
-                    && (!isIp || squareBracketsCount == 0) && (curlyBracketsCount == 0)) {
+            } else if ((!isIp || squareBracketsCount == 0) && (curlyBracketsCount == 0)
+                    && (delimiters != null && delimiters.indexOf(c) >= 0)) {
                 return sb.length() == 0 ? null : sb.toString();
             } else {
                 sb.append(c);