@schlosna 2017-06-09
While doing some performance analysis on a Jersey 2.25.1 based application, noticed that there where some oddities in `UriParser.parseComponent` where it was boxing `char` to `Character` and which then goes through less efficient `StringBuilder.append(Object)` path that does some unnecessary array copies & string creation rather than just appending `char` to buffer.

Here's an example backtrace snippet I saw:
```
  java.lang.String.valueOf(String.java:3008)
  java.lang.Character.toString(Character.java:4637)
  java.lang.String.valueOf(String.java:2994)
  java.lang.StringBuilder.append(StringBuilder.java:131)
  org.glassfish.jersey.uri.internal.UriParser.parseComponent(UriParser.java:126)
  org.glassfish.jersey.uri.internal.UriParser.parseComponentWithIP(UriParser.java:80)
  org.glassfish.jersey.uri.internal.UriParser.parseAuthority(UriParser.java:211)
  org.glassfish.jersey.uri.internal.UriParser.parseHierarchicalUri(UriParser.java:197)
  org.glassfish.jersey.uri.internal.UriParser.parse(UriParser.java:175)
  org.glassfish.jersey.uri.internal.JerseyUriBuilder.uri(JerseyUriBuilder.java:193)
  org.glassfish.jersey.uri.internal.JerseyUriBuilder.uri(JerseyUriBuilder.java:72)
  javax.ws.rs.core.UriBuilder.fromUri(UriBuilder.java:119)
  org.glassfish.jersey.servlet.ServletContainer.service(ServletContainer.java:298)
  org.glassfish.jersey.servlet.ServletContainer.service(ServletContainer.java:228)
```