@pavelbucek 2017-06-12
merged to the master branch.

Thanks!